% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amount_owed_after_n_months.r
\name{amount_owed_after_n_months}
\alias{amount_owed_after_n_months}
\title{"amount_owed_after_n_months": Calculates the outstanding mortgage principle after a set amount of months}
\usage{
amount_owed_after_n_months(annual_interest_rate, term, principal,
  month_of_interest, alternate_payment = NULL)
}
\arguments{
\item{month_of_interest}{The value of the month that you wish to know the outstanding principle}

\item{alternate_payment}{(OPTIONAL) A value for an alternate monthly payment used to pay off the principal - this is used for overpayments}
}
\value{
outstanding_owed The outstanding principle at the end of the supplied month of interest
}
\description{
Calculates the outstanding mortgage principle after a set amount of months.

This function calculates mortgage estimates. You should seek professional advice before making financial decisions.
}
\examples{
# amount_owed_after_n_months( annual_interest_rate , term , principal , month_of_interest )

}
\references{
Formulas sourced from https://en.wikipedia.org/wiki/Mortgage_calculator
}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
