% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_schedule.r
\name{debt_schedule}
\alias{debt_schedule}
\title{"debt_schedule": Calculates the mortgage debt schedule}
\usage{
debt_schedule(annual_interest_rate, term, principal, starting_date = NULL,
  alternate_payment = NULL)
}
\arguments{
\item{starting_date}{(OPTIONAL) If a starting date is supplied the schedule will autofill with years and dates}
}
\value{
debt_schedule A data frame containing the mortgage schedule
}
\description{
Calculates the mortgage debt schedule

This function calculates mortgage estimates. You should seek professional advice before making financial decisions.
}
\examples{
# debt_schedule( annual_interest_rate , term , principal , starting_date=NULL )

}
\references{
Formulas sourced from https://en.wikipedia.org/wiki/Mortgage_calculator
}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
