% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_schedule_with_intro_rate.r
\name{debt_schedule_with_intro_rate}
\alias{debt_schedule_with_intro_rate}
\title{"debt_schedule_with_intro_rate": Calculates the mortgage debt schedule}
\usage{
debt_schedule_with_intro_rate(annual_interest_rate, term, principal,
  introductory_interest_rate, introductory_term, starting_date = NULL,
  alternate_payment = NULL)
}
\arguments{
\item{introductory_interest_rate}{The introductory interest rate}

\item{introductory_term}{The introductory term in months}
}
\value{
debt_schedule A data frame containing the mortgage schedule
}
\description{
Calculates the mortgage debt schedule, when an introductory rate is available

This function calculates mortgage estimates. You should seek professional advice before making financial decisions.
}
\examples{
# debt_schedule_with_intro_rate( annual_interest_rate , term , principal , introductory_interest_rate , introductory_term , starting_date=NULL )

}
\references{
Formulas sourced from https://en.wikipedia.org/wiki/Mortgage_calculator
}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
