% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate_by_period.r
\name{interest_rate_by_period}
\alias{interest_rate_by_period}
\title{"interest_rate_by_period": Converts mortgage annual interest rates to periodic rates}
\usage{
interest_rate_by_period(annual_interest_rate, period = 12)
}
\arguments{
\item{annual_interest_rate}{The annual interest rate of the mortgage}

\item{period}{Optional and unused variable to change what period the annual interest rate is converted to. May be deprecated.}
}
\value{
periodic_interest_rate The interest rate for the period of note
}
\description{
Converts mortgage annual interest rates to periodic rates.

This function calculates mortgage estimates. You should seek professional advice before making financial decisions.
}
\examples{
# interest_rate_by_period( annual_interest_rate )
# interest_rate_by_period( annual_interest_rate , period = 12 )
# monthly_interest_rate <- interest_rate_by_period( annual_interest_rate )

}
\references{
Formulas sourced from https://en.wikipedia.org/wiki/Mortgage_calculator
}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
